scopes description changed string to list.
With 1.3.0, scopes declared as string struct. But if you follow this doc which is ``scopes (string): List of scopes to validate, separated by a space.``, you are gonna get ``json: cannot unmarshal string into Go struct field SignatureConfig.scopes of type []string`` error. So this statement is misleading. That why I changed it.
https://github.com/devopsfaith/krakend-jose/blob/f02ca595b00459c442f5f8a18891b1c113386db9/jws.go#L39